+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @khanacademy/perseus-editor
 
+## 6.4.1
+
+### Patch Changes
+
+-   [#1267](https://github.com/Khan/perseus/pull/1267) [`3977d361a`](https://github.com/Khan/perseus/commit/3977d361a91db647b79d499cc3f5b695af356f74) Thanks [@benchristel](https://github.com/benchristel)! - Internal: remove the `markings` property from the interactive graph state
+
+-   Updated dependencies [[`3977d361a`](https://github.com/Khan/perseus/commit/3977d361a91db647b79d499cc3f5b695af356f74), [`478398ff7`](https://github.com/Khan/perseus/commit/478398ff79cdd558256d628edce16cf14efe6f72)]:
+    -   @khanacademy/perseus@22.5.1
+
 ## 6.4.0
 
 ### Minor Changes